+++ pyobjc-core/Lib/objc/_compat.py
@@ -21,3 +21,19 @@ def setUseKVOForSetattr(value):
 def getUseKVOForSetattr():
     warnings.warn("Read objc.options.use_kvo instead", DeprecationWarning)
     return _options.use_kvo
+
+
+def _setClassExtender(value):
+    # XXX: Remove in PyObjC 9, internal function
+    warnings.warn("Set objc.options._class_extender instead", DeprecationWarning, 2)
+    _options._class_extender = value
+
+
+def allocateBuffer(length):
+    """Allocate a read/write buffer of memory of the given size."""
+    if not isinstance(length, int) or length <= 0:
+        raise TypeError("length must be a positive integer")
+
+    warnings.warn("Use bytearray instead", DeprecationWarning)
+
+    return bytearray(length)