+++ pyobjc-core/Modules/objc/OC_BuiltinPythonNumber.m
@@ -4,15 +4,21 @@ NS_ASSUME_NONNULL_BEGIN
 
 @implementation OC_BuiltinPythonNumber
 
-/* XXX: Never actually used because OC_PythonNumber uses
- * NSNumber for archiving for those instances that can
- * be represented by an NSNumber, resulting in the system
- * never using this method.
+/*
+ * This method is never actually used by the Foundation
+ * framework because ``-[OC_PythonNumber classForArchiver]``
+ * returns ``NSNumber`` for all values that are proxied
+ * using ``OC_BuiltinPythonNumber``.
+ *
+ * I'm leaving this method in just in case there is a
+ * valid usecase for this.
  */
+// LCOV_EXCL_START
 + (BOOL)supportsSecureCoding
 {
     return YES;
 }
+// LCOV_EXCL_STOP
 
 + (NSArray*)classFallbacksForKeyedArchiver
 {