+++ pyobjc-core/PyObjCTest/test_version_support.py
@@ -26,8 +26,19 @@ class TestVersionSupport(TestCase):
         self.assertTrue(objc.macos_available(version[0], version[1] - 1, version[2]))
         self.assertTrue(objc.macos_available(version[0], version[1] - 1, 0))
         self.assertTrue(objc.macos_available(version[0], version[1] - 1))
+        self.assertTrue(objc.macos_available(9, 0))
 
         if version[2]:
             self.assertTrue(
                 objc.macos_available(version[0], version[1], version[2] - 1)
             )
+
+        with self.assertRaisesRegex(
+            TypeError, r"function missing required argument 'major' \(pos 1\)"
+        ):
+            objc.macos_available()
+
+        with self.assertRaisesRegex(
+            TypeError, "'str' object cannot be interpreted as an integer"
+        ):
+            objc.macos_available("42", 0)